CUDA light implementation
Fix camera inversion

+++ light.h
@@ -10,13 +10,13 @@ public:
     static LightSource directional(float intensity, const glm::vec3 &dir, const glm::vec3 &color);
     static LightSource spherical(float intensity, const glm::vec3 &pos, const glm::vec3 &color);
 
-    glm::vec3 surface_color(const glm::dvec3 &surface_point,
-                            const glm::dvec3 &surface_norm,
+    glm::vec3 surface_color(const glm::vec3 &surface_point,
+                            const glm::vec3 &surface_norm,
                             const glm::vec3 &surface_color) const;
 
 #ifdef __NVCC__
-    __device__ glm::vec3 surface_color_cuda(const glm::dvec3 &surface_point,
-                            const glm::dvec3 &surface_norm,
+    __device__ glm::vec3 surface_color_cuda(const glm::vec3 &surface_point,
+                            const glm::vec3 &surface_norm,
                             const glm::vec3 &surface_color) const;
 #endif
